+++ src/diffusers/pipelines/__init__.py
@@ -34,6 +34,7 @@ if is_torch_available() and is_transformers_available():
         VersatileDiffusionTextToImagePipeline,
     )
     from .vq_diffusion import VQDiffusionPipeline
+    from .paint_by_example import PaintByExamplePipeline
 
 if is_transformers_available() and is_onnx_available():
     from .stable_diffusion import (
